Add tokenizer for gencode to allow spaces in lists.
Tokens must be separated by spaces.

If a token contains a ", `, or ' it may contain spaces within the quotes.

If a token contains a (, {, or [ it may contain spaces within the parens.

Thus, the string "foo(a, b, c)" is treated as one token, as is also
the string "{ hello(a, b, c); bar(); return ['x', 'y z'] }"