Add temporary "drawn" and "deck" panel displays.
Use this to show drawn cards while choosing cards for events that
draw X cards, play one, and put the rest on top / bottom.

Use stack of "draw" arrays in VM to handle nested events.
NOTE: This is not necessary now, but will be if we change the Winning Moves
event to play its event immediately.